Split the unit tests classes into separate handler/middleware tests classes
