+++ src/Owin.Security.OAuth.Introspection/OAuthIntrospectionMiddleware.cs
@@ -5,6 +5,7 @@
  */
 
 using System;
+using System.Collections.Generic;
 using System.Net.Http;
 using JetBrains.Annotations;
 using Microsoft.AspNetCore.DataProtection;
@@ -20,7 +21,7 @@ namespace Owin.Security.OAuth.Introspection {
     public class OAuthIntrospectionMiddleware : AuthenticationMiddleware<OAuthIntrospectionOptions> {
         public OAuthIntrospectionMiddleware(
             [NotNull] OwinMiddleware next,
-            [NotNull] IAppBuilder app,
+            [NotNull] IDictionary<string, object> properties,
             [NotNull] OAuthIntrospectionOptions options)
             : base(next, options) {
             if (string.IsNullOrEmpty(options.Authority) &&
@@ -38,14 +39,13 @@ namespace Owin.Security.OAuth.Introspection {
             }
 
             if (options.DataProtectionProvider == null) {
-                // Try to use the application name provided by
-                // the OWIN host as the application discriminator.
-                var discriminator = new AppProperties(app.Properties).AppName;
-
-                // When an application discriminator cannot be resolved from
-                // the OWIN host properties, generate a temporary identifier.
+                // Use the application name provided by the OWIN host as the Data Protection discriminator.
+                // If the application name cannot be resolved, throw an invalid operation exception.
+                var discriminator = new AppProperties(properties).AppName;
                 if (string.IsNullOrEmpty(discriminator)) {
-                    discriminator = Guid.NewGuid().ToString();
+                    throw new InvalidOperationException("The application name cannot be resolved from the OWIN application builder. " +
+                                                        "Consider manually setting the 'DataProtectionProvider' property in the " +
+                                                        "options using 'DataProtectionProvider.Create([unique application name])'.");
                 }
 
                 options.DataProtectionProvider = DataProtectionProvider.Create(discriminator);
